LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-expected.txt


PASS Popover focus test: default behavior - popover is not focused
FAIL Popover button click focus test: default behavior - popover is not focused assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: default behavior - popover is not focused
PASS Popover focus test: autofocus popover
FAIL Popover button click focus test: autofocus popover assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>



<div popover="" data-test="autofocus empty popo...
PASS Popover corner cases test: autofocus popover
PASS Popover focus test: autofocus empty popover
FAIL Popover button click focus test: autofocus empty popover assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>





<div popover="" data-test="autofocus popover ...
PASS Popover corner cases test: autofocus empty popover
PASS Popover focus test: autofocus popover with button
FAIL Popover button click focus test: autofocus popover with button assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>







<div popover="" data-test="autofocus child"...
PASS Popover corner cases test: autofocus popover with button
PASS Popover focus test: autofocus child
FAIL Popover button click focus test: autofocus child assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>









<div popover="" data-test="autofocus on t...
PASS Popover corner cases test: autofocus child
PASS Popover focus test: autofocus on tabindex=0 element
FAIL Popover button click focus test: autofocus on tabindex=0 element assert_false: clicking button should hide the popover expected false got true
FAIL Popover corner cases test: autofocus on tabindex=0 element assert_false: popover should start out hidden expected false got true
PASS Popover focus test: autofocus multiple children
FAIL Popover button click focus test: autofocus multiple children assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>













<div popover="" autofocus="" tabindex...
PASS Popover corner cases test: autofocus multiple children
PASS Popover focus test: autofocus popover and multiple autofocus children
FAIL Popover button click focus test: autofocus popover and multiple autofocus children assert_equals: Focus should go to unrelated button on light dismiss expected Element node <button></button> but got Element node <body>















<style>
  [popover] {
    border: 2...
PASS Popover corner cases test: autofocus popover and multiple autofocus children
